refactor: move export function to loadTools into navbar (#992)
Move `loadTools()` and helper functions from `tools.js` to
`loadTools.js` to allow reuse from within future `toolset.js`, which
will use similar logic.

Slightly modify `loadTools()` to accept a toolset name to allow
filtering. When called from `tools.js` it will take an empty string to
maintain existing behavior (fetch all tools).
